page loading with ajax and added hash route
