+++ spring-lens-ui/src/main/java/com/sdlcpro/springlens/ui/SpringLensUiAutoConfiguration.java
@@ -19,6 +19,7 @@ public class SpringLensUiAutoConfiguration implements WebMvcConfigurer {
 
     @Override
     public void addViewControllers(ViewControllerRegistry registry) {
+        registry.addRedirectViewController("/spring-lens", "/spring-lens/ui/index.html");
         registry.addRedirectViewController("/spring-lens/ui", "/spring-lens/ui/index.html");
     }
 }
