+++ spring-lens-ui/src/main/resources/META-INF/spring-lens/ui/assets/bean-definitions.json
