--- spring-lens-ui/src/main/resources/META-INF/spring-lens/ui/pages/bean.html
@@ -1,119 +1,0 @@
-<header class="page-header">
-    <div class="page-title">
-        <h1>Beans</h1>
-        <div class="crumbs">SpringLens &nbsp; &gt; &nbsp;Bean&nbsp; &gt; &nbsp;Definitions</div>
-    </div>
-
-    <div class="top-controls">
-        <button class="range"><span data-icon="clock"><svg viewBox="0 0 24 24" fill="none" stroke="currentColor">
-                    <circle cx="12" cy="12" r="9"></circle>
-                    <path d="M12 7v5l4 2"></path>
-                </svg></span>Last 15 minutes<span data-icon="chevron"><svg viewBox="0 0 24 24" fill="none"
-                    stroke="currentColor">
-                    <path d="m7 10 5 5 5-5"></path>
-                </svg></span></button>
-        <button class="icon-button" aria-label="Refresh"><span data-icon="refresh"><svg viewBox="0 0 24 24" fill="none"
-                    stroke="currentColor">
-                    <path d="M20 6v5h-5"></path>
-                    <path d="M4 18v-5h5"></path>
-                    <path d="M18 9a7 7 0 0 0-11.7-2.7L4 8.5"></path>
-                    <path d="M6 15a7 7 0 0 0 11.7 2.7L20 15.5"></path>
-                </svg></span></button>
-    </div>
-</header>
-<section class="module-layout">
-    <div class="module-main">
-        <section class="bean-metrics">
-            <article class="card module-card">
-                <div class="metric-top"><span class="mini-icon purple"
-                        style="color:#6d43e8;background:rgba(109,67,232,.12)"><svg viewBox="0 0 24 24" fill="none"
-                            stroke="currentColor">
-                            <path d="m12 3 8 4.5v9L12 21l-8-4.5v-9L12 3Z"></path>
-                            <path d="M4.5 8 12 12l7.5-4M12 21v-9"></path>
-                            <circle cx="12" cy="12" r="1.5"></circle>
-                        </svg></span>
-                    <h3>Total Bean Definitions</h3>
-                </div><strong>1,244</strong>
-                <div class="bar-list">
-                    <div class="bar-row"><span>Main</span><i style="--w:39%;--c:#6d43e8"></i></div>
-                    <div class="bar-row"><span>Auth</span><i style="--w:22%;--c:#2d9cdb"></i></div>
-                    <div class="bar-row"><span>Shared</span><i style="--w:48%;--c:#22b65f"></i></div>
-                </div>
-
-            </article>
-            <article class="card module-card">
-                <span class="mini-icon green"><i data-lucide="table-columns-split"></i></span>
-
-                <h3>Scope Distribution</h3>
-                <div class="donut-wrap" style="height:150px;gap:18px">
-                    <div class="donut" style="width:96px;height:96px"></div>
-                    <div class="legend"><span><i class="status-dot" style="color:#6d43e8"></i>Singleton
-                            34%<strong></strong></span><span><i class="status-dot" style="color:#2d9cdb"></i>Prototype
-                            10%<strong></strong></span><span><i class="status-dot" style="color:#22b65f"></i>Session
-                            8%<strong></strong></span><span><i class="status-dot" style="color:#ff9f1a"></i>Request
-                            23%<strong></strong></span></div>
-                </div>
-            </article>
-            <article class="card module-card">
-                <span class="mini-icon yellow"><i data-lucide="users"></i></span>
-                <h3>Roles Distribution</h3>
-                <div class="donut-wrap" style="height:150px;gap:18px">
-                    <div class="donut" style="width:96px;height:96px"></div>
-                    <div class="legend"><span><i class="status-dot" style="color:#6d43e8">
-
-                    </i>App 34%<strong></strong></span><span><i class="status-dot" style="color:#2d9cdb"></i>Support 10%<strong></strong></span><span><i class="status-dot" style="color:#22b65f"></i>Infra 8%<strong></strong></span>
-                    </div>
-                </div>
-            </article>
-            <article class="card module-card">
-                <div class="donut-wrap" style="height:150px;gap:18px">
-                    <div id="chart-container"></div>
-                </div>
-            </article>
-        </section>
-        <section class="datatable-for-bean"> </section>
-        <article class="card table-card">
-            <div class="filters-row">
-                <select class="filter-pill" name="method" id="method" class="">
-                    <option value="">All Methods</option>
-                    <option value="GET">GET</option>
-                    <option value="POST">POST</option>
-                    <option value="PUT">PUT</option>
-                    <option value="PATCH">PATCH</option>
-                    <option value="DELETE">DELETE</option>
-                </select>
-                <select class="filter-pill" name="status" id="status">
-                    <option value="">All Status</option>
-                    <option value="200">200</option>
-                    <option value="400">400</option>
-                    <option value="500">500</option>
-                    <option value="201">201</option>
-                    <option value="204">204</option>
-                </select>
-                <select class="filter-pill" name="services" id="services">
-                    <option value="">All Services</option>
-                    <option value="auth">Auth</option>
-                    <option value="products">Products</option>
-                    <option value="orders">Orders</option>
-                    <option value="cart">Cart</option>
-                    <option value="billing">Billing</option>
-                </select>
-            </div>
-            <div class="">
-                <table class="" id="tableCards">
-                    <thead>
-                        <tr>
-                            <th>Method</th>
-                            <th>Endpoint</th>
-                            <th>Status</th>
-                            <th>Response Time</th>
-                            <th>IP Address</th>
-                            <th>Time</th>
-                            <th>Action</th>
-                        </tr>
-                    </thead>
-                </table>
-            </div>
-        </article>
-    </div>
-</section>
